--- frontend/src/Edit.css
@@ -1,13 +1,0 @@
-.edit.container-fluid {
-    justify-content: center;
-    display: flex;
-}
-
-.edit.container-fluid .edit-form {
-    padding: .75rem 0;
-}
-
-.submit-container {
-    display: flex;
-    justify-content: center;
-}
